Add multi user support

+++ src/Api/Trakt/Cache/User/Movie/Rating/Repository.php
@@ -12,26 +12,27 @@ class Repository
     {
     }
 
-    public function clearCache() : void
+    public function clearCache(int $userId) : void
     {
-        $this->dbConnection->executeQuery('DELETE FROM `cache_trakt_user_movie_rating`');
+        $this->dbConnection->executeQuery('DELETE FROM `cache_trakt_user_movie_rating` WHERE user_id = ?', [$userId]);
     }
 
-    public function create(TraktId $traktId, int $rating, DateTime $ratedAt) : void
+    public function create(int $userId, TraktId $traktId, int $rating, DateTime $ratedAt) : void
     {
         $this->dbConnection->insert(
             'cache_trakt_user_movie_rating',
             [
                 'trakt_id' => $traktId->asInt(),
+                'user_id' => $userId,
                 'rating' => $rating,
                 'rated_at' => (string)$ratedAt,
             ]
         );
     }
 
-    public function findByTraktId(TraktId $traktId) : ?int
+    public function findByTraktId(int $userId, TraktId $traktId) : ?int
     {
-        $data = $this->dbConnection->fetchOne('SELECT rating FROM `cache_trakt_user_movie_rating` WHERE trakt_id = ?', [$traktId->asInt()]);
+        $data = $this->dbConnection->fetchOne('SELECT rating FROM `cache_trakt_user_movie_rating` WHERE trakt_id = ? AND user_id = ?', [$traktId->asInt(), $userId]);
 
         return $data === false ? null : (int)$data;
     }
